Refactored line protocol parsing to its own package
