+++ lineproto/influx_line_protocol_test.go
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Rackspace US, Inc.
+ * Copyright 2020 Rackspace US, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package agents
+package lineproto
 
 import (
 	"github.com/racker/salus-telemetry-protocol/telemetry_edge"
